+++ cx-portal/src/components/pages/ServiceAdminBoardDetail/index.tsx
@@ -34,7 +34,7 @@ import {
 import { useTranslation } from 'react-i18next'
 import { PAGES } from 'types/Constants'
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
-import { InputLabel } from '@mui/material'
+import { InputLabel, Grid } from '@mui/material'
 import { download } from 'utils/downloadUtils'
 
 export default function ServiceAdminBoardDetail() {
@@ -79,6 +79,22 @@ export default function ServiceAdminBoardDetail() {
     } catch (error) {
       console.error(error, 'ERROR WHILE FETCHING DOCUMENT')
     }
+  }
+
+  const getTechUserData = (data: string[] | null) => {
+    return data && data?.length > 0 ? (
+      data?.map((role: string) => (
+        <Typography variant="subtitle2" key={role}>
+          * {role}
+        </Typography>
+      ))
+    ) : (
+      <Typography variant="caption2" className="not-available">
+        {t(
+          'adminboardDetail.technicalUserSetup.noTechnicalUserProfilesAvailable'
+        )}
+      </Typography>
+    )
   }
 
   return (
@@ -163,6 +179,28 @@ export default function ServiceAdminBoardDetail() {
                   </span>
                 </InputLabel>
               ))}
+          </div>
+
+          <div style={{ marginBottom: '60px' }}>
+            <Typography variant="h4">
+              {t('adminboardDetail.technicalUserSetup.heading')}
+            </Typography>
+            <Typography
+              variant="body2"
+              style={{
+                marginTop: '20px',
+              }}
+            >
+              {t('adminboardDetail.technicalUserSetup.message')}
+            </Typography>
+            <Grid container spacing={2} sx={{ margin: '30px 0' }}>
+              <Grid item xs={12} style={{ padding: '0px' }}>
+                {serviceData.technicalUserProfile &&
+                  getTechUserData(
+                    Object.values(serviceData?.technicalUserProfile)[0]
+                  )}
+              </Grid>
+            </Grid>
           </div>
 
           <div className="adminboard-provider">