+++ cx-portal/src/components/shared/basic/ReleaseProcess/components/CommonValidateAndPublish.tsx
@@ -250,6 +250,20 @@ export default function CommonValidateAndPublish({
     }
   }
 
+  const getTechUserData = (data: string[] | null) => {
+    return data && data?.length > 0 ? (
+      data?.map((role: string) => (
+        <Typography variant="subtitle2" key={role}>
+          * {role}
+        </Typography>
+      ))
+    ) : (
+      <Typography variant="caption2" className="not-available">
+        {t('global.errors.noTechnicalUserProfilesAvailable')}
+      </Typography>
+    )
+  }
+
   return (
     <div className="validate-and-publish">
       <ReleaseStepHeader
@@ -519,6 +533,22 @@ export default function CommonValidateAndPublish({
             )}
           </>
         )}
+
+        {statusData?.technicalUserProfile &&
+          Object.values(statusData?.technicalUserProfile) && (
+            <>
+              <Divider className="verify-validate-form-divider" />
+              <Typography variant="h4" sx={{ mb: 4 }}>
+                {t('content.adminboardDetail.technicalUserSetup.heading')}
+              </Typography>
+              <Typography variant="body2" className="form-field">
+                {t('content.adminboardDetail.technicalUserSetup.message')}
+              </Typography>
+              {getTechUserData(
+                Object.values(statusData?.technicalUserProfile)[0]
+              )}
+            </>
+          )}
 
         <Divider className="verify-validate-form-divider" />
         <Typography variant="h4" sx={{ mb: 4 }}>